source/bin/requirement-test.txt
urllib3==1.25.10
pytest==6.1.2
pytest-cov==2.10.1
pyyaml==5.3
    